Initial Goap support
This is just a really rough sketch of the goap system that I plan to implement.
For now, I have just managed to get the system some what working, but there seem to be many bugs.
I am also considering changing the goap system to work without MonoBehaviour, but I don't know yet.